1.5

+++ app/Http/Controllers/API/TileController.php
@@ -15,52 +15,42 @@ class TileController extends Controller
             "data" => ["tiles" => Tile::where('user_id', auth()->id())->get()]
         ]);
     }
-/*
+
     public function store()
     {
+        // Store tile using tile url
+
         $tile = new Tile;
         $tile->user_id = auth()->id();
         $tile->name = request('name');
 
-        if (request()->hasFile('icon')) {
-            $path = request()->file('icon')->store('tiles', 'public');
-            $tile->icon = '/storage/' . $path;
-        }
+        if (request()->hasFile('image')) {
+            $path = request()->file('image')->store('tiles', 'public');
+            $tile->url = '/storage/' . $path;
+
+            list($width, $height) = getimagesize(public_path($tile->url));
+            $tile->bounds = [[0, 0],[$height, $width]];
+        } else $tile->url = request('url');
 
         $tile->save();
 
         return response()->json([
             "status" => "success",
-            "data" => ["tiles" => Tile::where('user_id', auth()->id())->get()]
-        ], 200);
-    }
-
-    public function update($id)
-    {
-        $tile = Tile::find($id);
-
-        if (request('name')) $tile->name = request('name');
-        if (request()->hasFile('icon')) {
-            // Remove old icon
-            if (File::exists(public_path($tile->icon))) {
-                $icon_basename = basename($tile->icon);
-                if ($icon_basename != 'biology.png' || $icon_basename != 'computer_science.png' || $icon_basename != 'custom.png' || $icon_basename != 'geography.png' || $icon_basename != 'history.png')
-                    File::delete(public_path($tile->icon));
-            }
-            // Save new icon
-            $path = request()->file('icon')->store('tiles', 'public');
-            $tile->icon = '/storage/' . $path;
-        }
-        $tile->save();
-
-        return response()->json([
-            "status" => "success",
-            "data" => ["tiles" => Tile::where('user_id', auth()->id())->get()]
-        ], 200);
+            "data" => [
+                "tiles" => Tile::where('user_id', auth()->id())->get()
+            ]
+        ]);
     }
 
     public function destroy($id)
     {
+
+        if (Tile::find($id)->user_id !== auth()->id()) {
+            return response()->json([
+                "status" => "success",
+                "data" => ["id" => "Access denied"]
+            ], 403);
+        }
         // Get all tiles for current user.
         Tile::destroy($id);
 
@@ -68,5 +58,5 @@ class TileController extends Controller
             "status" => "success",
             "data" => ["tiles" => Tile::where('user_id', auth()->id())->get()]
         ], 200);
-    }*/
+    }
 }
